Move the binary operator factory functions to the `computation_constructing_utils` module.
PiperOrigin-RevId: 259560717
