+++ tensorflow_federated/python/core/impl/computation_constructing_utils_test.py
@@ -34,6 +34,7 @@ from tensorflow_federated.python.core.impl import computation_wrapper_instances
 from tensorflow_federated.python.core.impl import context_stack_impl
 from tensorflow_federated.python.core.impl import intrinsic_defs
 from tensorflow_federated.python.core.impl import placement_literals
+from tensorflow_federated.python.core.impl import type_constructors
 from tensorflow_federated.python.core.impl import type_utils
 from tensorflow_federated.python.core.impl import value_impl
 
@@ -2756,6 +2757,206 @@ class TensorFlowConstantTest(absltest.TestCase):
     self.assertTrue(
         np.array_equal(executable_noarg_zero()[0].b,
                        np.ones([2, 2], dtype=np.float32)))
+
+
+class BinaryOperatorTest(absltest.TestCase):
+
+  def test_apply_op_raises_on_none(self):
+    with self.assertRaisesRegex(TypeError, 'ComputationBuildingBlock'):
+      computation_constructing_utils.apply_binary_operator_with_upcast(
+          None, tf.multiply)
+
+  def test_construct_op_raises_on_none_operator(self):
+    with self.assertRaisesRegex(TypeError, 'found non-callable'):
+      computation_constructing_utils.create_binary_operator_with_upcast(
+          tf.int32, None)
+
+  def test_raises_incompatible_tuple_and_tensor(self):
+    bad_type_ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType([[tf.int32, tf.int32], tf.float32],
+                                        placement_literals.CLIENTS))
+    with self.assertRaisesRegex(TypeError, 'incompatible with upcasted'):
+      computation_constructing_utils.apply_binary_operator_with_upcast(
+          bad_type_ref, tf.multiply)
+    with self.assertRaisesRegex(TypeError, 'incompatible with upcasted'):
+      computation_constructing_utils.create_binary_operator_with_upcast(
+          bad_type_ref.type_signature.member, tf.multiply)
+
+  def test_raises_non_callable_op(self):
+    bad_type_ref = computation_building_blocks.Reference(
+        'x', [tf.float32, tf.float32])
+    with self.assertRaisesRegex(TypeError, 'non-callable'):
+      computation_constructing_utils.apply_binary_operator_with_upcast(
+          bad_type_ref, tf.constant(0))
+    with self.assertRaisesRegex(TypeError, 'non-callable'):
+      computation_constructing_utils.create_binary_operator_with_upcast(
+          bad_type_ref, tf.constant(0))
+
+  def test_raises_tuple_and_nonscalar_tensor(self):
+    bad_type_ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType(
+            [[tf.int32, tf.int32],
+             computation_types.TensorType(tf.float32, [2])],
+            placement_literals.CLIENTS))
+    with self.assertRaisesRegex(TypeError, 'incompatible with upcasted'):
+      computation_constructing_utils.apply_binary_operator_with_upcast(
+          bad_type_ref, tf.multiply)
+    with self.assertRaisesRegex(TypeError, 'incompatible with upcasted'):
+      computation_constructing_utils.create_binary_operator_with_upcast(
+          bad_type_ref.type_signature.member, tf.multiply)
+
+  def test_raises_tuple_scalar_multiplied_by_nonscalar(self):
+    bad_type_ref = computation_building_blocks.Reference(
+        'x', [tf.int32, computation_types.TensorType(tf.float32, [2])])
+    with self.assertRaisesRegex(TypeError, 'incompatible with upcasted'):
+      computation_constructing_utils.apply_binary_operator_with_upcast(
+          bad_type_ref, tf.multiply)
+    with self.assertRaisesRegex(TypeError, 'incompatible with upcasted'):
+      computation_constructing_utils.create_binary_operator_with_upcast(
+          bad_type_ref.type_signature, tf.multiply)
+
+  def test_construct_generic_raises_federated_type(self):
+    bad_type = computation_types.FederatedType(
+        [[tf.int32, tf.int32],
+         computation_types.TensorType(tf.float32, [2])],
+        placement_literals.CLIENTS)
+    with self.assertRaisesRegex(TypeError, 'argument that is not a two-tuple'):
+      computation_constructing_utils.create_binary_operator_with_upcast(
+          bad_type, tf.multiply)
+
+  def test_apply_integer_type_signature(self):
+    ref = computation_building_blocks.Reference('x', [tf.int32, tf.int32])
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.multiply)
+    self.assertEqual(multiplied.type_signature,
+                     computation_types.to_type(tf.int32))
+
+  def test_construct_integer_type_signature(self):
+    ref = computation_building_blocks.Reference('x', [tf.int32, tf.int32])
+    multiplier = computation_constructing_utils.create_binary_operator_with_upcast(
+        ref.type_signature, tf.multiply)
+    self.assertEqual(
+        multiplier.type_signature,
+        type_constructors.binary_op(computation_types.to_type(tf.int32)))
+
+  def test_multiply_federated_integer_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType([tf.int32, tf.int32],
+                                        placement_literals.CLIENTS))
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.multiply)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.FederatedType(tf.int32, placement_literals.CLIENTS))
+
+  def test_divide_federated_float_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType([tf.float32, tf.float32],
+                                        placement_literals.CLIENTS))
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.multiply)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.FederatedType(tf.float32, placement_literals.CLIENTS))
+
+  def test_multiply_federated_unnamed_tuple_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType(
+            [[tf.int32, tf.float32], [tf.int32, tf.float32]],
+            placement_literals.CLIENTS))
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.multiply)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.FederatedType([tf.int32, tf.float32],
+                                        placement_literals.CLIENTS))
+
+  def test_multiply_federated_named_tuple_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType(
+            [[('a', tf.int32),
+              ('b', tf.float32)], [('a', tf.int32), ('b', tf.float32)]],
+            placement_literals.CLIENTS))
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.multiply)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.FederatedType([('a', tf.int32), ('b', tf.float32)],
+                                        placement_literals.CLIENTS))
+
+  def test_divide_federated_named_tuple_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType(
+            [[('a', tf.int32),
+              ('b', tf.float32)], [('a', tf.int32), ('b', tf.float32)]],
+            placement_literals.CLIENTS))
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.divide)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.FederatedType([('a', tf.float64), ('b', tf.float32)],
+                                        placement_literals.CLIENTS))
+
+  def test_multiply_federated_named_tuple_with_scalar_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType([[('a', tf.float32),
+                                          ('b', tf.float32)], tf.float32],
+                                        placement_literals.CLIENTS))
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.multiply)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.FederatedType([('a', tf.float32), ('b', tf.float32)],
+                                        placement_literals.CLIENTS))
+
+  def test_multiply_named_tuple_with_scalar_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x', [[('a', tf.float32), ('b', tf.float32)], tf.float32])
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.multiply)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.NamedTupleType([('a', tf.float32),
+                                          ('b', tf.float32)]))
+
+  def test_construct_multiply_op_named_tuple_with_scalar_type_signature(self):
+    type_spec = computation_types.to_type([[('a', tf.float32),
+                                            ('b', tf.float32)], tf.float32])
+    multiplier = computation_constructing_utils.create_binary_operator_with_upcast(
+        type_spec, tf.multiply)
+    expected_function_type = computation_types.FunctionType(
+        type_spec, type_spec[0])
+    self.assertEqual(multiplier.type_signature, expected_function_type)
+
+  def test_construct_divide_op_named_tuple_with_scalar_type_signature(self):
+    type_spec = computation_types.to_type([[('a', tf.float32),
+                                            ('b', tf.float32)], tf.float32])
+    multiplier = computation_constructing_utils.create_binary_operator_with_upcast(
+        type_spec, tf.divide)
+    expected_function_type = computation_types.FunctionType(
+        type_spec, type_spec[0])
+    self.assertEqual(multiplier.type_signature, expected_function_type)
+
+  def test_divide_federated_named_tuple_with_scalar_type_signature(self):
+    ref = computation_building_blocks.Reference(
+        'x',
+        computation_types.FederatedType([[('a', tf.float32),
+                                          ('b', tf.float32)], tf.float32],
+                                        placement_literals.CLIENTS))
+    multiplied = computation_constructing_utils.apply_binary_operator_with_upcast(
+        ref, tf.divide)
+    self.assertEqual(
+        multiplied.type_signature,
+        computation_types.FederatedType([('a', tf.float32), ('b', tf.float32)],
+                                        placement_literals.CLIENTS))
 
 
 if __name__ == '__main__':
